requirements.txt
fastapi==0.75.2
uvicorn==0.17.6
sqlalchemy==1.4.35
psycopg2-binary==2.9.3
python-keycloak==0.27.0
alembic==1.7.6
